Updated solution for prismic-hook, sitemap.txt and link-cache. Still only 2-levels.
