+++ index.js
@@ -15,14 +15,8 @@ var stores = require('./lib/middleware/stores')
 var prismic = require('./lib/middleware/prismic')
 var lang = require('./lib/middleware/lang')
 var navigation = require('./lib/middleware/navigation')
-var Db = require('node-cache')
 
 var server = new Koa()
-
-/**
- *  Cache for navigation links
- */
-server.context.db = new Db()
 
 /**
  * Compile and serve assets on demand during development
