+++ lib/app/stores/index.js
@@ -2,8 +2,8 @@
 module.exports = function defaultStore (ctx) {
   return {
     version: process.env.npm_package_version,
-    db: ctx.db,
     lang: process.env.AL_LANG,
+    url: process.env.AL_URL,
     locale: null,
     error: null,
     ref: null,
