+++ lib/assets/sitemap.txt
@@ -1,0 +1,117 @@
+http://localhost:3000/en/astrids-characters/characters---mardie
+http://localhost:3000/sv/karaktarerna/karaktarer---rasmus-pa-luffen
+http://localhost:3000/sv/karaktarerna/lotta-pa-brakmakargatan
+http://localhost:3000/en/astrids-characters/lotta-and-her-family-live-in-a-yellow-house-on-potmaker
+http://localhost:3000/sv/pippi-langstrump/pippi-langstrump-efraim-langstrump
+http://localhost:3000/sv/pippi-langstrump/pippi-langstrump-herr-nilsson
+http://localhost:3000/sv/pippi-langstrump/pippi-langstrump-kling-klang
+http://localhost:3000/sv/foraldrarna/foraldrarna
+http://localhost:3000/sv/karaktarerna/pippi-langstrump
+http://localhost:3000/sv/pippi-langstrump/pippi-langstrump-tommy-annika
+http://localhost:3000/sv/pippi-langstrump/pippi-roligheter
+http://localhost:3000/sv/nils-karlsson-pyssling-flyttar-in
+http://localhost:3000/sv/boklista
+http://localhost:3000/sv/pippi-langstrump/skola-bibliotek
+http://localhost:3000/sv/i-ett-alldeles-vanligt-hus-pa-en-alldeles-vanlig-gata
+http://localhost:3000/sv/raben--sjogren5
+http://localhost:3000/sv/mio-min-mio
+http://localhost:3000/sv/britt-mari-lattar-sitt-hjarta-ar-astrid-lindgrens-debutbo
+http://localhost:3000/sv/vargsangen
+http://localhost:3000/sv/astrid-lindgren
+http://localhost:3000/sv/i-den-tredje-fristaende-kati-boken-gar-farden-till
+http://localhost:3000/sv/raben--sjogren6
+http://localhost:3000/sv/pippi-langstrump/pippi-langstrump-lilla-gubben
+http://localhost:3000/sv/astrid-lindgren/utvalda-artal
+http://localhost:3000/sv/karaktarerna/mio
+http://localhost:3000/sv/verken
+http://localhost:3000/sv/foretaget
+http://localhost:3000/sv/astrid-lindgren/katla
+http://localhost:3000/sv/pippi-langstrump/historien-bakom
+http://localhost:3000/sv/citaten
+http://localhost:3000/sv/kalle-blomkvist-och-rasmus
+http://localhost:3000/sv/masterdetektiven-kalle-blomkvist-ar-bara-tretton-ar
+http://localhost:3000/sv/raben--sjogren3
+http://localhost:3000/sv/ingrid-vang-nyman
+http://localhost:3000/sv/aret-ar-1950-strax-efter-kriget-och-visst-ar-det-ett
+http://localhost:3000/sv/raben--sjogren4
+http://localhost:3000/sv/norstedts
+http://localhost:3000/sv/boken-om-pippi-langstrump
+http://localhost:3000/sv/ilon-wikland-birgitta-nordenskjold
+http://localhost:3000/sv/pa-kaptensgatan-i-en-liten-trevlig-lya-bor-kati.-hon
+http://localhost:3000/sv/kajsa-kavat
+http://localhost:3000/sv/jag-vill-inte-ga-och-lagga-mig
+http://localhost:3000/sv/kanner-du-pippi-langstrump
+http://localhost:3000/sv/astrid-lindgren-skrev-manga-manus-till-teaterpjaser.
+http://localhost:3000/sv/norstedts1
+http://localhost:3000/sv/den-andra-boken-om-barnen-i-bullerbyn-som-har-roligt
+http://localhost:3000/sv/karaktarerna/masterdetektiven
+http://localhost:3000/sv/pippi-langstrump/sanger-filmer
+http://localhost:3000/sv/ansostestsida
+http://localhost:3000/sv/bara-roligt-i-bullerbyn
+http://localhost:3000/sv/raben--sjogren2
+http://localhost:3000/sv/astrid-lindgren/skolan
+http://localhost:3000/sv/astrid-lindgren/syskonen
+http://localhost:3000/de/tysk-test
+http://localhost:3000/sv/astrid-lindgren/barndomen
+http://localhost:3000/sv/bullerbyn/historien-bakom-bullerbyn
+http://localhost:3000/sv/idas-sommarvisa
+http://localhost:3000/sv/du-kare-lille-snickerbo
+http://localhost:3000/sv/verken/sangerna
+http://localhost:3000/sv/pilutta-visan
+http://localhost:3000/sv/har-kommer-pippi-langstrump
+http://localhost:3000/sv/luffarvisan
+http://localhost:3000/sv/verken/film-tv
+http://localhost:3000/sv/var-pa-saltkrakan
+http://localhost:3000/sv/platt-plitt-plutt-visan
+http://localhost:3000/sv/falukorvsvisan
+http://localhost:3000/sv/foretaget/foretaget-bocker
+http://localhost:3000/sv/verken/illustratorerna
+http://localhost:3000/sv/verken/teatern
+http://localhost:3000/sv/varldens-basta-karlsson
+http://localhost:3000/sv/foretaget/foretaget-licensprodukterna
+http://localhost:3000/sv/foretaget/foretaget-teatern
+http://localhost:3000/sv/foretaget/foretaget-kontakt
+http://localhost:3000/sv/alla-ska-sova-for-nu-ar-det-natt
+http://localhost:3000/sv/historien-bakom/anjas-test
+http://localhost:3000/en/astrids-characters/the-children-of-noisy-village
+http://localhost:3000/sv/foretaget/foretaget-film
+http://localhost:3000/en/astrids-characters
+http://localhost:3000/sv/foretaget/foretaget-forlaget
+http://localhost:3000/sv/madicken/historien-bakom-madicken
+http://localhost:3000/sv/karaktarerna/bullerbyn
+http://localhost:3000/sv/filmlista
+http://localhost:3000/sv/pippi-langstrump-gar-ombord
+http://localhost:3000/sv/karaktarerna/karlsson-pa-taket
+http://localhost:3000/sv/tema
+http://localhost:3000/sv/kerstin-och-jag
+http://localhost:3000/sv/karaktarerna
+http://localhost:3000/sv/pippi-langstrump
+http://localhost:3000/sv/pippi-langstrump-i-soderhavet
+http://localhost:3000/sv/karaktarerna/madicken
+http://localhost:3000/sv/karaktarerna/broderna-lejonhjarta
+http://localhost:3000/sv/karaktarerna/emil-i-lonneberga
+http://localhost:3000/sv/astrid-lindgren/ungdomen
+http://localhost:3000/sv/valdigt-bra-bild-pa-lotta-pa-brakmakargatan
+http://localhost:3000/sv/pressnyheter
+http://localhost:3000/sv/astrids-bilder
+http://localhost:3000/sv/visste-du-att
+http://localhost:3000/sv/verken/kort-test
+http://localhost:3000/sv/besoksmal
+http://localhost:3000/sv/traffa-alla-karaktarer
+http://localhost:3000/sv/emil-i-lonnebega
+http://localhost:3000/sv/emil-i-lonneberga1
+http://localhost:3000/sv/astrid-lindgren/efter-pensionen
+http://localhost:3000/sv/astrid-lindgren/alderdomen-och-doden
+http://localhost:3000/sv/astrid-lindgren/vuxenlivet1
+http://localhost:3000/sv/astrid-lindgren/minnet-av-astrid-lindgren
+http://localhost:3000/sv/kontakt
+http://localhost:3000/en/a-short-biography
+http://localhost:3000/de/der-mensch-astrid-lindgren
+http://localhost:3000/sv/an-lever-emil-i-lonneberga1
+http://localhost:3000/sv/madicken-1
+http://localhost:3000/sv/nya-hyss-av-emil-i-lonneberga
+http://localhost:3000/sv/madicken-och-junibackens-pims
+http://localhost:3000/sv/emil-i-lonneberga
+http://localhost:3000/sv/broderna-lejonhjarta
+http://localhost:3000/sv/ronja-rovardotter
+http://localhost:3000/sv/emil-i-lonneberga/emil-hyss
