+++ lib/components/card/index.js
@@ -99,7 +99,7 @@ function Body (body, { title, titleLink }, isLargerText, hasBg, hasTitle, layout
 }
 
 function MoreLink ({href, label}, state) {
-  let link = linkResolver(state.db, {lang: state.lang, ...href})
+  let link = linkResolver({lang: state.lang, ...href})
   return html`
     <a class="Card-moreLink" href="${link}">${label}</a>
   `
