+++ lib/components/footer/index.js
@@ -14,7 +14,7 @@ let Footer = (data, ctx) => {
   if (ctx && ctx.state && ctx.state.params && Array.isArray(ctx.state.params)) {
     breadcrums = html`
       <div class="Footer-breadcrums">${ctx.state.params.map((item) => {
-        return html`<span>/</span><a href="${resolve(ctx.state.db, {
+        return html`<span>/</span><a href="${resolve({
           link_type: 'Document',
           uid: item,
           type: 'page',
@@ -44,7 +44,7 @@ let Footer = (data, ctx) => {
             <span class="Footer-content-linksblock-title">${item.primary.title}</span>
             ${item.items.map((item) => {
               return html`
-                <a class="Footer-content-linksblock-link" href="${resolve(ctx.state.db, item.link)}">${item.title}</a>
+                <a class="Footer-content-linksblock-link" href="${resolve(item.link)}">${item.title}</a>
               `
             })}
           </div>
