+++ lib/components/slices/characters/index.js
@@ -36,7 +36,7 @@ function shelf (slice, state, container, ctx) {
       type: 'character',
       isRoundCover: true,
       cover: html`
-        <a href="${linkResolver(ctx ? ctx.state.db : state.db, {
+        <a href="${linkResolver({
           uid: item.uid,
           lang: state.lang,
           link_type: 'Document',
@@ -52,7 +52,7 @@ function shelf (slice, state, container, ctx) {
       `,
       info: html`
         <div class="Text Text--smaller">
-          <p><a href="${linkResolver(ctx ? ctx.state.db : state.db, {
+          <p><a href="${linkResolver({
             uid: item.uid,
             lang: state.lang,
             link_type: 'Document',
