+++ lib/components/slices/shelf/index.js
@@ -21,7 +21,7 @@ function shelf (slice, state, container) {
   books.forEach((item) => {
     if (!item || !item.data) return
     var title = item.data.book_title[0].text
-    const href = linkResolver(state.db, item)
+    const href = linkResolver(item)
     const profile = image(item.data.cover, ['qw-small', 'qw-medium', 'qw-large', 'qw-xlarge'])
 
     items.push({
@@ -57,7 +57,7 @@ function shelf (slice, state, container) {
         ${data.titel ? html`<div class="Display-title">${asElement(data.titel)}</div>` : ''}
     `,
     footer: html`
-      ${data.more_label ? html`<div class="Text"><p><a href="${linkResolver(state.db, data.more_link)}">${data.more_label}</a></p></div>` : ''}
+      ${data.more_label ? html`<div class="Text"><p><a href="${linkResolver(data.more_link)}">${data.more_label}</a></p></div>` : ''}
     `,
     body: Display(state, items)
   })
