+++ lib/components/subnavigation/index.js
@@ -42,7 +42,7 @@ function subNavigation (state, context) {
 
 function subNavItem (link) {
   if (hasLabel(link)) {
-    const path = resolve(ctx.state.db, link)
+    const path = resolve(link)
 
     const classes = className('SubNav-link', {
       'SubNav-link--active': link.isActive
@@ -58,7 +58,7 @@ function subNavItem (link) {
 
 function subNavCharacterItem (link) {
   if (hasLabel(link)) {
-    const path = resolve(ctx.state.db, link)
+    const path = resolve(link)
 
     const classes = className('SubNav-link', {
       'SubNav-link--active': link.isActive
