+++ lib/components/text/serializer.js
@@ -67,7 +67,7 @@ function serialize (options = {}) {
  */
 
 function resolve (link) {
-  if (link.link_type === 'Document') return linkResolver(null, link)
+  if (link.link_type === 'Document') return linkResolver(link)
   const { hostname, pathname, hash } = url.parse(link.url)
   if (!hostname) return (pathname || '') + (hash || '')
   if (hostname.indexOf('.') === -1) return '/' + hostname + (pathname || '')
