+++ lib/render.js
@@ -7,8 +7,6 @@ const Page404 = require('./router/404')
  * @param {next} function async component
  */
 module.exports = async (ctx, next) => {
-  ctx.state.db = ctx.db
-
   try {
     await next()
