+++ lib/router/index.js
@@ -7,7 +7,7 @@
 const compose = require('koa-compose')
 
 const routes = [
-  require('./sitemap'),
+  require('./hook'),
   require('./preview'),
   require('./home'),
   require('./newsdesk'),
