+++ lib/router/preview.js
@@ -8,7 +8,7 @@ const linkResolver = require('../resolve.js')
 router.get(process.env.PRISMIC_PREVIEW, async function (ctx, next) {
   const token = ctx.request.query.token
   const url = await ctx.prismic.previewSession(token, async (res) => {
-    return linkResolver(ctx.db, res)
+    return linkResolver(res)
   }, '/')
   console.log('[Preview] token:', token, 'url:', url)
   ctx.cookies.set(Prismic.previewCookie, token, { maxAge: 30 * 60 * 1000, path: '/', httpOnly: false })
